Redesigned Login & SignUp Page (Responsive)
