+++ client/src/app/auth/login/page.tsx
@@ -1,6 +1,10 @@
 "use client";
 import { useEffect, useState } from "react";
 import Link from "next/link";
+import LoginSvg from "@/public/assets/Login.svg";
+import { useThemeContext } from '../../layout';
+import Image from "next/image";
+
 import useFetch from "@/src/hooks/useFetch";
 import { AuthService } from "@/src/services/AuthService";
 import LoadingSpinner from "@/src/components/Loaders/LoadingSpinner";
@@ -10,6 +14,8 @@ import useProfileStore from "@/src/store/profileStore";
 import { useRouter } from "next/navigation";
 import { ProfileService } from "@/src/services/ProfileService";
 import Header from "@/src/components/Headers/Header";
+import Navbar from "@/src/components/Headers/Navbar";
+import Footer from "@/src/components/Footer/Footer";
 
 const Login = () => {
   const router = useRouter();
@@ -72,64 +78,70 @@ const Login = () => {
       );
     }
   };
-
+  const { theme, toggleTheme } = useThemeContext();
   return (
-    <div data-testid="login" className=" ">
-      <div className="flex flex-col">
-        <Header />
-        <div className="flex flex-col items-center justify-center">
-          <div className="bg-white p-5 rounded shadow m-2 flex flex-col w-4/5 lg:w-1/3 mt-16">
-            <h1 className="text-3xl self-center">Login</h1>
-
-            <input
-              className="border rounded m-2 h-12 p-2 shadow-sm"
-              type="email"
-              value={email}
-              onChange={(e) => setEmail(e.target.value)}
-              placeholder="Enter your email"
-            />
-            <input
-              className="border rounded m-2 h-12 p-2 shadow-sm"
-              type="password"
-              value={password}
-              onChange={(e) => setPassword(e.target.value)}
-              placeholder="Enter your password"
-            />
-
-            <button
-              onClick={onHandleSubmit}
-              disabled={isLoginLoading}
-              className="w-full bg-blue-400 border rounded m-2 h-12 p-2 shadow-sm"
-            >
-              Submit
-            </button>
-
-            {isLoginLoading ? (
-              <div className="w-full flex justify-center">
-                <LoadingSpinner />
+    <>
+    <Navbar landingPage={false} />
+    <div className={`flex w-screen  flex-wrap ${theme === 'light' ? 'text-slate-800 ' : 'text-white  border-t-2 border-white bg-[#12141D]'}`}>
+      <div className="flex w-full  h-auto flex-col md:w-1/2">
+        <div className="my-3 mx-auto flex flex-col justify-center px-6 pt-8 md:justify-start lg:w-[28rem]">
+          <p className="text-center text-3xl font-bold md:leading-tight md:text-left md:text-5xl">
+            Welcome back <br />
+            to <span className="text-blue-600">KeySafe.</span>
+          </p>
+          <p className="mt-6 text-center font-medium md:text-left">Sign in to your account below.</p>
+          <form className="flex flex-col items-stretch pt-3 md:pt-8" onSubmit={onHandleSubmit}>
+            <div className="flex flex-col pt-4">
+              <div className="relative flex overflow-hidden rounded-md border border-gray-500 transition focus-within:border-blue-600">
+                <input
+                  type="email"
+                  id="login-email"
+                  className={`w-full flex-shrink appearance-none ${theme === 'light' ? 'bg-[white] text-gray-700 ' : 'bg-[#12141D] text-white '} border-gray-500 py-2 px-4 text-base  placeholder-gray-400 focus:outline-none`}
+                  placeholder="Email"
+                  value={email}
+                  onChange={(e) => setEmail(e.target.value)}
+                />
               </div>
-            ) : null}
-
-            <div className="text-red-500 font-semibold w-full flex justify-center">
-              {errorMessage ? errorMessage : ""}
             </div>
-          </div>
-          <div className="flex flex-col items-center mt-12">
-            <p className="">
-              Don&apos;t have an account?{" "}
-              <span>
-                <Link
-                  className="text-teal-600 hover:underline"
-                  href="/auth/register"
-                >
-                  Register
-                </Link>
-              </span>
+            <div className="mb-4 flex flex-col pt-4">
+              <div className="relative flex overflow-hidden rounded-md border border-gray-500 transition focus-within:border-blue-600">
+                <input
+                  type="password"
+                  id="login-password"
+                  className={`w-full flex-shrink appearance-none ${theme === 'light' ? 'bg-[white]  text-gray-700' : 'bg-[#12141D] text-white'} border-gray-500 py-2 px-4 text-base  placeholder-gray-400 focus:outline-none`}
+                  placeholder="Password"
+                  value={password}
+                  onChange={(e) => setPassword(e.target.value)}
+                />
+              </div>
+            </div>
+            <a href="#" className="mb-6 text-center text-sm font-medium text-gray-600 md:text-left">Forgot password?</a>
+            <button
+              disabled={isLoginLoading}
+              type="submit"
+              className="rounded-lg bg-blue-600 px-4 py-2 text-center text-base font-semibold text-white shadow-md outline-none ring-blue-500 ring-offset-2 transition hover:bg-blue-700 focus:ring-2 md:w-32"
+            >
+              Sign in
+            </button>
+          </form>
+          <div className="py-12 text-center">
+            <p className="text-gray-600">
+              Don't have an account?
+              <Link href="/auth/register" className={`whitespace-nowrap px-1 font-semibold ${theme === 'light' ? 'text-gray-900' : 'text-white'} underline underline-offset-4`}>Sign up for free.</Link>
             </p>
           </div>
         </div>
       </div>
+      <div className="relative hidden px-10 select-none bg-blue-600 bg-gradient-to-br md:block md:w-1/2">
+        <div className="pt-9 py-4 px-8 text-white xl:w-[40rem]">
+          <p className="mb-6 mt-1 text-3xl font-semibold leading-10">Never Forget Your Passwords Again!<span className="abg-white whitespace-nowrap py-2 text-cyan-300"> Safe & Secure</span>.</p>
+          <p className="mb-4">Managing passwords can be tough, but with KeySafe, you don't have to worry about forgetting them anymore.</p>
+        </div>
+        <Image alt='login illustration' src={LoginSvg} className="ml-8 w-11/12 max-w-lg py-5 rounded-lg object-cover" />
+      </div>
     </div>
+    <Footer />
+  </>
   );
 };
