+++ client/src/app/auth/register/page.tsx
@@ -1,4 +1,10 @@
 "use client";
+import LoginSvg from "@/public/assets/Login.svg";
+import { useThemeContext } from '../../layout';
+import Image from "next/image";
+import Navbar from "@/src/components/Headers/Navbar";
+import Footer from "@/src/components/Footer/Footer";
+
 import { useEffect, useState } from "react";
 import LoadingSpinner from "@/src/components/Loaders/LoadingSpinner";
 import { useRouter } from "next/navigation";
@@ -9,9 +15,9 @@ import { setCookie } from "cookies-next";
 import useAuthStore from "@/src/store/authStore";
 import useProfileStore from "@/src/store/profileStore";
 import { ProfileService } from "@/src/services/ProfileService";
-import Header from "@/src/components/Headers/Header";
 
-const Register = () => {
+export default function Register() {
+
   const router = useRouter();
   const setAuthToken = useAuthStore((state) => state.setAuthToken);
   const setProfile = useProfileStore((state) => state.setProfileDetails);
@@ -87,70 +93,86 @@ const Register = () => {
     }
   };
 
+  const { theme, toggleTheme } = useThemeContext();
   return (
-    <div data-testid="register" className="">
-      <div className="flex flex-col">
-        <Header />
-        <div className="flex flex-col items-center justify-center">
-          <div className="bg-white p-5  rounded shadow m-2 flex flex-col w-4/5 lg:w-1/3 mt-16">
-            <h1 className="text-3xl self-center">Register</h1>
-            <input
-              className="border rounded m-2 h-12 p-2 shadow-sm"
-              type="text"
-              value={name}
-              onChange={(e) => setName(e.target.value)}
-              placeholder="Enter your name"
-            />
-            <input
-              className="border rounded m-2 h-12 p-2 shadow-sm"
-              type="email"
-              value={email}
-              onChange={(e) => setEmail(e.target.value)}
-              placeholder="Enter your email"
-            />
-            <input
-              className="border rounded m-2 h-12 p-2 shadow-sm"
-              type="password"
-              value={password}
-              onChange={(e) => setPassword(e.target.value)}
-              placeholder="Enter your password"
-            />
+    <>
+    <Navbar landingPage={false} />
+    <div className={`flex w-screen  flex-wrap ${theme === 'light' ? 'text-slate-800 ' : 'border-t-2 border-white text-white bg-[#12141D]'}`}>
+      <div className="relative hidden h- select-none bg-blue-600 bg-gradient-to-br md:block md:w-1/2">
+        <div className="pt-9 py-4 px-8 text-white xl:w-[40rem]">
+          <p className="mb-6 mt-1 text-3xl font-semibold leading-10">Never Forget Your Passwords Again!<span className="abg-white whitespace-nowrap py-2 text-cyan-300"> Safe & Secure</span>.</p>
+          <p className="mb-4">Tired of forgetting your passwords? Say goodbye to sticky notes and random paper scraps with Key Safe.</p>
+        </div>
+        <Image alt='login illustration' src={LoginSvg} className="ml-8 w-11/12 max-w-lg py-5 rounded-lg object-cover" />
+      </div>
+      <div className="flex w-full h-auto flex-col md:w-1/2">
+        <div className="my-3 mx-auto flex flex-col justify-center px-6 pt-8 md:justify-start lg:w-[28rem]">
+          <p className="text-center text-3xl font-bold md:leading-tight md:text-left md:text-5xl">
+            Welcome to <br />
+            <span className="text-blue-600">KeySafe.</span>
+          </p>
+          <p className="mt-6 text-center font-medium md:text-left">Create your account below.</p>
+          <form className="flex flex-col items-stretch pt-3 md:pt-8" >
 
+          <div className="flex flex-col pt-4">
+              <div className="relative flex overflow-hidden rounded-md border border-gray-500 transition focus-within:border-blue-600">
+                <input
+                  type="text"
+                  id="register-name"
+                  className={`w-full flex-shrink appearance-none ${theme === 'light' ? 'bg-[white] text-gray-700 ' : 'bg-[#12141D] text-white '} border-gray-500 py-2 px-4 text-base  placeholder-gray-400 focus:outline-none`}
+                  placeholder="Name"
+                  value={name}
+                  onChange={(e) => setName(e.target.value)}
+                />
+              </div>
+            
+            </div>
+
+            <div className="flex flex-col pt-4">
+              <div className="relative flex overflow-hidden rounded-md border border-gray-500 transition focus-within:border-blue-600">
+                <input
+                  type="email"
+                  id="register-email"
+                  className={`w-full flex-shrink appearance-none ${theme === 'light' ? 'bg-[white] text-gray-700 ' : 'bg-[#12141D] text-white '} border-gray-500 py-2 px-4 text-base  placeholder-gray-400 focus:outline-none`}
+                  placeholder="Email"
+                  value={email}
+                  onChange={(e) => setEmail(e.target.value)}
+                />
+              </div>
+            
+            </div>
+            <div className="mb-4 flex flex-col pt-4">
+              <div className="relative flex overflow-hidden rounded-md border border-gray-500 transition focus-within:border-blue-600">
+                <input
+                  type="password"
+                  id="register-password"
+                  className={`w-full flex-shrink appearance-none ${theme === 'light' ? 'bg-[white]  text-gray-700' : 'bg-[#12141D] text-white'} border-gray-500 py-2 px-4 text-base  placeholder-gray-400 focus:outline-none`}
+                  placeholder="Password"
+                  value={password}
+                  onChange={(e) => setPassword(e.target.value)}
+                  />
+              </div>
+            </div>
+            <a href="#" className="mb-6 text-center text-sm font-medium text-gray-600 md:text-left">Forgot password?</a>
             <button
               onClick={onHandleSubmit}
               disabled={isRegisterLoading}
-              className="w-full bg-blue-400 border rounded m-2 h-12 p-2 shadow-sm"
+              type="submit"
+              className="rounded-lg bg-blue-600 px-4 py-2 text-center text-base font-semibold text-white shadow-md outline-none ring-blue-500 ring-offset-2 transition hover:bg-blue-700 focus:ring-2 md:w-32"
             >
-              Submit
+              Sign Up
             </button>
-
-            {isRegisterLoading ? (
-              <div className="w-full flex justify-center">
-                <LoadingSpinner />
-              </div>
-            ) : null}
-
-            <div className="text-red-500 font-semibold w-full flex justify-center">
-              {errorMessage ? errorMessage : ""}
-            </div>
-          </div>
-          <div className="flex flex-col items-center mt-12">
-            <p className="">
-              Already have an account?{" "}
-              <span>
-                <Link
-                  className="text-teal-600 hover:underline"
-                  href="/auth/login"
-                >
-                  Login
-                </Link>
-              </span>
+          </form>
+          <div className="py-12 text-center">
+            <p className="text-gray-600">
+              Already have an account?
+              <Link href="/auth/login" className={`whitespace-nowrap px-1 font-semibold ${theme === 'light' ? 'text-gray-900' : 'text-white'} underline underline-offset-4`}>Login</Link>
             </p>
           </div>
         </div>
       </div>
     </div>
+    <Footer />
+  </>
   );
 };
-
-export default Register;
